# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ATS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: roleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: submitCandidate
      tags: [candidates]
      summary: Submit a candidate (Vendor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCandidateRequest"
      responses:
        "201":
          description: Candidate submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "409":
          description: Duplicate candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates/{id}:
    get:
      tags: [candidates]
      operationId: getCandidate
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Candidate detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '404':
          description: Not found
  /candidates/{id}/status:
    patch:
      operationId: updateCandidateStatus
      tags: [candidates]
      summary: Update candidate status (Client HR)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateStatusRequest"
      responses:
        "200":
          description: Candidate status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
  /candidates/{id}/notes:
    get:
      tags: [notes]
      operationId: listCandidateNotes
      summary: List notes for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateNote'
    post:
      tags: [notes]
      operationId: addCandidateNote
      summary: Add a note to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddNoteRequest'
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateNote'
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        expectedSalary:
          type: number
          nullable: true
        status:
          type: string
          enum: [submitted, screening, interview, offer, hired, rejected]
        roleId:
          type: integer
        roleTitle:
          type: string
        vendorCompanyId:
          type: integer
        vendorCompanyName:
          type: string
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cvUrl:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
      required:
        - id
        - firstName
        - lastName
        - email
        - status
        - roleId
        - roleTitle
        - vendorCompanyId
        - vendorCompanyName
        - submittedAt
        - updatedAt
    SubmitCandidateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        expectedSalary:
          type: number
        roleId:
          type: integer
        cvUrl:
          type: string
        tags:
          type: string
      required:
        - firstName
        - lastName
        - email
        - roleId
    UpdateCandidateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, screening, interview, offer, hired, rejected]
      required:
        - status
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
...
    CandidateNote:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - userId
        - authorName
        - content
        - createdAt
    AddNoteRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...