# lib/api-zod/src/generated/types/updateCandidateStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCandidateStatusRequestStatus } from "./updateCandidateStatusRequestStatus";

export interface UpdateCandidateStatusRequest {
  status: UpdateCandidateStatusRequestStatus;
}